Refactor source structure
