--- src/background/tldts.js
@@ -1,2 +1,0 @@
-import * as tldts from 'tldts-experimental';
-globalThis.tldts = tldts;
