+++ src/pages/options/ghostery-options.js
@@ -12,28 +12,7 @@
 import { html, define, store } from 'hybrids';
 import { t } from '/vendor/@whotracksme/ui/src/i18n.js';
 
-const Options = {
-  trackerWheelDisabled: false,
-  wtmSerpReport: true,
-  [store.connect]: {
-    async get() {
-      const storage = await chrome.storage.local.get(['options']);
-      return storage.options || {};
-    },
-    async set(_, options, keys) {
-      const prevOptions = await this.get();
-      const nextOptions = {
-        ...prevOptions,
-        ...Object.fromEntries(keys.map((key) => [key, options[key]])),
-      };
-
-      chrome.storage.local.set({ options: nextOptions });
-      chrome.runtime.sendMessage({ action: 'updateOptions' });
-
-      return options;
-    },
-  },
-};
+import Options from '/store/options.js';
 
 export default define({
   tag: 'ghostery-options',
