+++ src/pages/panel/index.js
@@ -11,11 +11,11 @@
 
 import '/vendor/@whotracksme/ui/src/components/category-bullet.js';
 import '/vendor/@whotracksme/ui/src/components/panel-header.js';
+import '/vendor/@whotracksme/ui/src/components/wtm-stats.js';
 
-import './components/layout.js';
+import '/components/layout.js';
+import '/components/page-load.js';
+import '/components/toggle-switch.js';
+import '/components/category-with-trackers.js';
 
-import './components/page-load.js';
-import './components/toggle-switch.js';
-import './components/category-with-trackers.js';
-
-import './app.js';
+import './views/app.js';
