--- src/pages/panel/store/tracker.js
@@ -1,17 +1,0 @@
-/**
- * Ghostery Browser Extension
- * https://www.ghostery.com/
- *
- * Copyright 2017-present Ghostery GmbH. All rights reserved.
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0
- */
-
-export default {
-  id: true,
-  name: '',
-  company_id: 'unknown',
-  category: 'unknown',
-};
