+++ src/pages/panel/views/detailed.js
@@ -11,10 +11,11 @@
 
 import { html, define, store, router } from 'hybrids';
 import { t } from '/vendor/@whotracksme/ui/src/i18n.js';
+
 import WTMTrackerWheel from '/vendor/@whotracksme/ui/src/tracker-wheel.js';
 import { chevronLeft } from '/vendor/@whotracksme/ui/src/components/icons.js';
 
-import Stats from '../store/stats.js';
+import Stats from '/store/stats.js';
 
 export default define({
   tag: 'gh-panel-detailed-view',
