+++ src/pages/panel/views/simple.js
@@ -12,60 +12,70 @@
 import { html, define, store, router } from 'hybrids';
 import { t } from '/vendor/@whotracksme/ui/src/i18n.js';
 
-import '/vendor/@whotracksme/ui/src/components/wtm-stats.js';
 import {
   externalLink,
   chevronRight,
 } from '/vendor/@whotracksme/ui/src/components/icons.js';
 
-import { rulesetIds } from '../utils/rulesets.js';
 import sites from '../../../rule_resources/sites.json';
 
-import Stats from '../store/stats.js';
-import Settings from '../store/settings.js';
+import Stats from '/store/stats.js';
+import Rulesets, { rulesetIds } from '../../../store/rulesets.js';
 
 import Detailed from './detailed.js';
+
+function toggleRuleset(ruleset) {
+  return (host) => {
+    store.set(host.rulesets, {
+      [ruleset]: !host.rulesets[ruleset],
+    });
+  };
+}
 
 export default define({
   [router.connect]: { stack: [Detailed] },
   tag: 'gh-panel-simple-view',
-  settings: store(Settings),
+  rulesets: store(Rulesets),
   stats: store(Stats),
-  render: ({ settings, stats }) => html`
+  render: ({ rulesets, stats }) => html`
     <h1>${t('privacy_protection')}</h1>
 
     <section class="toggles">
-      ${store.ready(settings) &&
+      ${store.ready(rulesets) &&
       rulesetIds.map(
-        (toggle) =>
+        (ruleset) =>
           html`<gh-panel-toggle-switch
-            toggle=${toggle}
+            name="${ruleset}"
+            disabled="${!rulesets[ruleset]}"
+            onclick=${toggleRuleset(ruleset)}
           ></gh-panel-toggle-switch>`,
       )}
+      ${store.error(rulesets) && html`<p>${store.error(rulesets)}</p>`}
     </section>
 
     ${store.ready(stats) &&
-    html` <wtm-stats categories=${stats.categories}></wtm-stats> `}
+    html`
+      <wtm-stats categories=${stats.categories}></wtm-stats>
+      <section class="buttons">
+        <span>
+          ${store.ready(stats) &&
+          sites.indexOf(stats.domain) > -1 &&
+          html`
+            <a
+              href="https://www.whotracks.me/websites/${stats.domain}.html"
+              target="_blank"
+            >
+              ${t('statistical_report')} ${externalLink}
+            </a>
+          `}
+        </span>
+        <a href="${router.url(Detailed)}">
+          ${t('detailed_view')} ${chevronRight}
+        </a>
+      </section>
 
-    <section class="buttons">
-      <span>
-        ${store.ready(stats) &&
-        sites.indexOf(stats.domain) > -1 &&
-        html`
-          <a
-            href="https://www.whotracks.me/websites/${stats.domain}.html"
-            target="_blank"
-          >
-            ${t('statistical_report')} ${externalLink}
-          </a>
-        `}
-      </span>
-      <a href="${router.url(Detailed)}">
-        ${t('detailed_view')} ${chevronRight}
-      </a>
-    </section>
-
-    <gh-panel-page-load></gh-panel-page-load>
+      <gh-panel-page-load loadTime="${stats.loadTime}"></gh-panel-page-load>
+    `}
   `.css`
     h1 {
       color: var(--black);
